+++ x/evidence/keeper/abci.go
@@ -1,10 +1,9 @@
-package evidence
+package keeper
 
 import (
 	"fmt"
 	"time"
 
-	"cosmossdk.io/x/evidence/keeper"
 	"cosmossdk.io/x/evidence/types"
 	abci "github.com/cometbft/cometbft/abci/types"
 
@@ -14,7 +13,7 @@ import (
 
 // BeginBlocker iterates through and handles any newly discovered evidence of
 // misbehavior submitted by CometBFT. Currently, only equivocation is handled.
-func BeginBlocker(ctx sdk.Context, req abci.RequestBeginBlock, k keeper.Keeper) {
+func (k *Keeper) BeginBlocker(ctx sdk.Context, req abci.RequestBeginBlock) {
 	defer telemetry.ModuleMeasureSince(types.ModuleName, time.Now(), telemetry.MetricKeyBeginBlocker)
 
 	for _, tmEvidence := range req.ByzantineValidators {
@@ -23,7 +22,7 @@ func BeginBlocker(ctx sdk.Context, req abci.RequestBeginBlock, k keeper.Keeper) 
 		// premeditation. So for now we agree to treat them in the same way.
 		case abci.MisbehaviorType_DUPLICATE_VOTE, abci.MisbehaviorType_LIGHT_CLIENT_ATTACK:
 			evidence := types.FromABCIEvidence(tmEvidence)
-			k.HandleEquivocationEvidence(ctx, evidence.(*types.Equivocation))
+			k.handleEquivocationEvidence(ctx, evidence.(*types.Equivocation))
 
 		default:
 			k.Logger(ctx).Error(fmt.Sprintf("ignored unknown evidence type: %s", tmEvidence.Type))
